refactor: extract TrendsChartSection from TrendsView + ratchet thresholds (#80, pt 2c)
Final TrendsView decomposition. Moves the inline chart experience — hero
variants, the Swift Charts plot (draw-on reveal + drag-to-inspect crosshair),
moving-average legend, stats row, single-point and stepped-back-empty variants
— into TrendsChartSection, which owns the inline crosshair selection
(inlineSelectedDate) and reveal-animation state (chartRevealProgress). It
receives vm/goalVM/compare config + secondaryColor and reports expand / set-goal
/ manage-goal via closures.

TrendsView is now a ~290-line orchestrator (state, sheets, onChange wiring,
the metric chip / range tabs / cold-start empty dispatch). The rangeTabs
no longer clears inlineSelectedDate directly — the chart clears its own
selection on timeRange change.

  TrendsView.swift: 1129 → 287 lines.

With TrendsView's giant type body gone, the #80 ratchet lands:
  - file_length 1700 → 1200 (now pinned by ScanEntryFlow, 1171)
  - type_body_length 1350 → 1000 (now pinned by ScanEntryFlow body, 993)

Verbatim move — no behavior change. Verified: make lint (0, --strict, with
the tightened thresholds), make build, make test (272 logic + 13 UI green),
and on-device (metric/range reveal, crosshair drag, stepper, compare,
single-point, stepped-back-empty, fullscreen).
